\name{ishpR-package}
\alias{ishpR-package}
\alias{ishpR}
\docType{package}
\title{\packageTitle{ishpR}}
\description{\packageDescription{ishpR}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{ishpR}
  \packageIndices{ishpR}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{ishpR}
  
  Maintainer: \packageMaintainer{ishpR}
}
\references{
Juan C. Laria, M. Carmen Aguilera-Morillo & Rosa E. Lillo (2019) An Iterative Sparse-Group Lasso, Journal of Computational and Graphical Statistics, DOI: 10.1080/10618600.2019.1573687
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
